requirements.txt
cohere==5.5.0
psycopg==3.1.19
python-dotenv==1.0.1
